requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aioimaplib==1.0.1
beautifulsoup4==4.12.2
aiohttp==3.9.1
requests==2.31.0
reportlab==4.0.7
pydantic==2.5.2
lxml==5.1.0
